requirements.txt
fastapi
uvicorn
faiss-cpu
pypdf
numpy
requests
python-multipart